Automatically move git status window to bottom with small height.
When the commit message window is closing, something in fugitive
conflicts with running vim.cmd("wincmd J") which causes the whole
commit to be cancelled.

If there is a way to only run this move and resize once when the
git status window is created, this would probably not be an issue.
Until then, defer_fn works to asynchronously move the window after
the conflict has passed.